fastapi/pyproject.toml
[project]
name = "dc-control-tower-agents"
version = "0.1.0"
description = "FastAPI agent runtime for DC Control Tower — debates, extractions, synthesis via Claude Opus + Sonnet."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "anthropic>=0.40.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sse-starlette>=2.1.3",
    "pypdf>=5.1.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
